pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = "GraphRAG Creation Pipeline"
authors = ["shavvimal <shavhugan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
environs = "^11.0.0"
datashaper = "^0.0.49"
# Vector Stores
azure-search-documents = "^11.4.0"
lancedb = "^0.11.0"
# Event Loops
uvloop = { version = "^0.19.0", markers = "platform_system != 'Windows'" }
nest-asyncio = { version = "^1.6.0", markers = "platform_system == 'Windows'" }
# Async IO
aiolimiter = "^1.1.0"
aiofiles = "^24.1.0"
# LLM
openai = "^1.37.1"
nltk = "3.8.1"
tiktoken = "^0.7.0"
# Data-Sci
numba = "0.60.0"
numpy = "^1.25.2"
graspologic = "^3.4.1"
networkx = "^3"
fastparquet = "^2024.2.0"
# 1.13.0 was a footgun
scipy = "1.12.0"
# Configuration
pyyaml = "^6.0.1"
pyaml-env = "^1.2.1"
python-dotenv = "^1.0.0"
# Network
tenacity = "^8.5.0"
swifter = "^1.4.0"
pydantic = "^2"
rich = "^13.6.0"
textual = "^0.74.0"
devtools = "^0.12.2"
typing-extensions = "^4.12.2"
#Azure
azure-storage-blob = "^12.19.0"
azure-identity = "^1.17.1"
json-repair = "^0.25.3"
asyncpg = "^0.29.0"
youtube-transcript-api = "^0.6.2"
motor = "^3.5.1"
google-api-python-client = "^2.140.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.6.0"
ipykernel = "^6.29.4"
jupyter = "^1.0.0"
nbconvert = "^7.16.3"
poethepoet = "^0.27.0"
pyright = "^1.1.371"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.4"
pytest-timeout = "^2.3.1"
ruff = "^0.5.2"
semversioner = "^2.0.3"
update-toml = "^0.2.1"
# poetry add yfiles_jupyter_graphs --group dev
yfiles-jupyter-graphs = "^1.7.3"

[build-system]
# A list of dependencies required to build your package, making this key mandatory
requires = ["poetry-core"]
# The Python object used to perform the build process
build-backend = "poetry.core.masonry.api"